Implement M8 Track Cost Integration with dynamic pricing
Implement dynamic track cost tracking based on Runner Framework's track_cost event with configurable multiplier. Replace fixed token costs with actual execution costs from Runner.

Key Changes:
- Database: Migrate token amounts from Integer to Float, allow negative balance down to -1000.0
- TrackCost model: Parse track_cost SSE event data from Runner Framework
- Cost multiplier: Configurable parameter (default 3.14) applied to raw costs
- CVService/ApplyService: Track cost during execution, charge after completion
- TokenService: Support float amounts, allow negative balance but block new executions
- API schemas: Update all token fields to float
- Documentation: Complete technical documentation for implementation

Database Migration:
- Migration 542d7e567590: Convert Integer to Float for token_balance, tokens_delta, balance_after
- Update check constraint to allow token_balance >= -1000.0

Service Changes:
- Added _track_cost instance variable to track execution cost
- Parse track_cost event in stream message handlers
- Apply cost_multiplier from settings (configurable via COST_MULTIPLIER env var)
- Round final cost to 2 decimal places
- Fallback to 0.0 if track_cost event not received
- Store cost metadata in transaction (cost_raw, cost_multiplier, cost_final)

User Experience:
- Display "Потрачено X токенов" after successful track execution
- Allow balance to go negative during charge
- Block new track execution when balance < 0

+++ backend/src/db/repositories/user_repository.py
@@ -68,7 +68,7 @@ class UserRepository:
     async def update_balance(
         self,
         user_id: int,
-        delta: int,
+        delta: float,
         expected_version: int,
     ) -> User:
         """Update token balance with optimistic locking.
